Review Tab with Custom Listview
Review Tab made with ListView using custom Adapter. Took 2 days- a lot of time. Swiping up ListView only doesn't scroll up the tabs bar and collapse the collapsingToolbar: it had to be done by dragging up the collapsingToolbar or the TabRibbon. To be fixed.
Custom NavigationDrawer to follow for now.